# packages/contracts/openapi/cinewatch-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "ExternalAvailabilityResponse": {
        "additionalProperties": false,
        "properties": {
          "media_type": {
            "enum": [
              "movie",
              "tv"
            ],
            "title": "Media Type",
            "type": "string"
          },
          "region": {
            "maxLength": 8,
            "minLength": 2,
            "title": "Region",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ExternalAvailabilitySource"
            },
            "maxItems": 200,
            "title": "Sources",
            "type": "array"
          },
          "tmdb_id": {
            "exclusiveMinimum": 0.0,
            "title": "Tmdb Id",
            "type": "integer"
          }
        },
        "required": [
          "media_type",
          "tmdb_id",
          "region"
        ],
        "title": "ExternalAvailabilityResponse",
        "type": "object"
      },
...
      "ExternalTrailer": {
        "additionalProperties": false,
        "properties": {
          "categories": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Categories",
            "type": "array"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "media_type": {
            "anyOf": [
              {
                "enum": [
                  "movie",
                  "tv"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "provider_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Provider Id",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "resource_title": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Title"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "tmdb_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Id"
          },
          "youtube_video_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Youtube Video Id",
            "type": "string"
          }
        },
        "required": [
          "provider_id",
          "youtube_video_id",
          "title"
        ],
        "title": "ExternalTrailer",
        "type": "object"
      },
      "ExternalTrailerLibraryResponse": {
        "additionalProperties": false,
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "total_pages": {
            "minimum": 0.0,
            "title": "Total Pages",
            "type": "integer"
          },
          "total_results": {
            "minimum": 0.0,
            "title": "Total Results",
            "type": "integer"
          },
          "trailers": {
            "items": {
              "$ref": "#/components/schemas/ExternalTrailer"
            },
            "maxItems": 50,
            "title": "Trailers",
            "type": "array"
          }
        },
        "required": [
          "page",
          "total_pages",
          "total_results"
        ],
        "title": "ExternalTrailerLibraryResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LyricsLookupResponse": {
        "additionalProperties": false,
        "properties": {
          "album": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "ambiguous": {
            "default": false,
            "title": "Ambiguous",
            "type": "boolean"
          },
          "artist": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Artist",
            "type": "string"
          },
          "candidate_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Candidate Count",
            "type": "integer"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/LyricsMatch"
            },
            "maxItems": 50,
            "title": "Candidates",
            "type": "array"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LyricsMatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          },
          "term": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "term",
          "artist",
          "resolved"
        ],
        "title": "LyricsLookupResponse",
        "type": "object"
      },
      "LyricsMatch": {
        "additionalProperties": false,
        "properties": {
          "album": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "album_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album Url"
          },
          "artist": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Artist",
            "type": "string"
          },
          "artist_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Url"
          },
          "song": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Song",
            "type": "string"
          },
          "song_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Song Url"
          }
        },
        "required": [
          "song",
          "artist"
        ],
        "title": "LyricsMatch",
        "type": "object"
      },
...
      "NewsResponse": {
        "additionalProperties": false,
        "properties": {
          "articles": {
            "items": {
              "$ref": "#/components/schemas/NewsArticle"
            },
            "maxItems": 20,
            "title": "Articles",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "query": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "total_results": {
            "minimum": 0.0,
            "title": "Total Results",
            "type": "integer"
          }
        },
        "required": [
          "query",
          "page",
          "page_size",
          "total_results"
        ],
        "title": "NewsResponse",
        "type": "object"
      },
...
      "ScriptLookupResponse": {
        "additionalProperties": false,
        "properties": {
          "ambiguous": {
            "default": false,
            "title": "Ambiguous",
            "type": "boolean"
          },
          "candidate_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Candidate Count",
            "type": "integer"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScriptMatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "requested_title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Requested Title",
            "type": "string"
          },
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          }
        },
        "required": [
          "requested_title",
          "resolved"
        ],
        "title": "ScriptLookupResponse",
        "type": "object"
      },
      "ScriptMatch": {
        "additionalProperties": false,
        "properties": {
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "excerpt": {
            "anyOf": [
              {
                "maxLength": 900,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Genres",
            "type": "array"
          },
          "has_script_text": {
            "default": false,
            "title": "Has Script Text",
            "type": "boolean"
          },
          "scene_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Count"
          },
          "script_format": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script Format"
          },
          "script_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script Url"
          },
          "source": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "word_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word Count"
          },
          "writers": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Writers",
            "type": "array"
          },
          "year": {
            "anyOf": [
              {
                "maximum": 2200.0,
                "minimum": 1800.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "title"
        ],
        "title": "ScriptMatch",
        "type": "object"
      },
      "SearchResponse": {
        "additionalProperties": false,
        "properties": {
          "query": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Query",
            "type": "string"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/SearchSuggestion"
            },
            "maxItems": 5,
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VoiceSampleStageRequest": {
        "additionalProperties": false,
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/VoiceSampleMetadata"
          },
          "wav_base64": {
            "maxLength": 20000000,
            "minLength": 16,
            "title": "Wav Base64",
            "type": "string"
          }
        },
        "required": [
          "metadata",
          "wav_base64"
        ],
        "title": "VoiceSampleStageRequest",
        "type": "object"
      },
      "VoiceSampleStageResponse": {
        "additionalProperties": false,
        "properties": {
          "retention_status": {
            "const": "LOCAL_PRIVATE_STAGING",
            "default": "LOCAL_PRIVATE_STAGING",
            "title": "Retention Status",
            "type": "string"
          },
          "sample_id": {
            "maxLength": 120,
            "minLength": 12,
            "title": "Sample Id",
            "type": "string"
          },
          "stored": {
            "title": "Stored",
            "type": "boolean"
          }
        },
        "required": [
          "sample_id",
          "stored"
        ],
        "title": "VoiceSampleStageResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "CineWatch TV V1 backend service boundary.",
    "title": "CineWatch TV API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/external/lyrics": {
      "get": {
        "operationId": "v1_external_lyrics",
        "parameters": [
          {
            "in": "query",
            "name": "term",
            "required": true,
            "schema": {
              "maxLength": 300,
              "minLength": 1,
              "title": "Term",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "artist",
            "required": true,
            "schema": {
              "maxLength": 300,
              "minLength": 1,
              "title": "Artist",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "album",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 300,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Album"
            }
          },
          {
            "in": "query",
            "name": "reference_url",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reference Url"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LyricsLookupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lyrics Lookup",
        "tags": [
          "external-test"
        ]
      }
    },
    "/api/v1/external/scripts": {
      "get": {
        "operationId": "v1_external_scripts",
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "required": true,
            "schema": {
              "maxLength": 300,
              "minLength": 1,
              "title": "Title",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2200,
                  "minimum": 1800,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptLookupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Script Lookup",
        "tags": [
          "external-test"
        ]
      }
    },
    "/api/v1/external/trailers": {
      "get": {
        "operationId": "v1_external_trailers",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "latest",
              "enum": [
                "latest",
                "trending"
              ],
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalTrailerLibraryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trailer Library",
        "tags": [
          "external-test"
        ]
      }
    },
    "/api/v1/external/where-to-watch": {
      "get": {
        "operationId": "v1_external_where_to_watch",
        "parameters": [
          {
            "in": "query",
            "name": "media_type",
            "required": true,
            "schema": {
              "enum": [
                "movie",
                "tv"
              ],
              "title": "Media Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tmdb_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Tmdb Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "default": "US",
              "maxLength": 8,
              "minLength": 2,
              "title": "Region",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalAvailabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Where To Watch",
        "tags": [
          "external-test"
        ]
      }
    },
...
    "/api/v1/news": {
      "get": {
        "operationId": "v1_news",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "film OR television",
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 20,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Live CineWatch entertainment news",
        "tags": [
          "news"
        ]
      }
    },
    "/api/v1/search": {
      "get": {
        "operationId": "v1_search",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Typed CineWatch entity search",
        "tags": [
          "search"
        ]
      }
    },
...
    "/api/v1/voice-lab/samples": {
      "post": {
        "operationId": "v1_voice_stage_sample",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceSampleStageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceSampleStageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage an explicitly accepted NexVox WAV sample",
        "tags": [
          "voice-lab"
        ]
      }
    },
...
  }
}